+++ CHANGELOG.md
@@ -25,11 +25,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Add MainContainer to organize sections and integrate inventory data
+- Adds a json file containing sample data from Airtable
 - Add forms for inventory and shopping list management
 - Add static components: ItemCard, InventorySection, Footer, Header, NavMenu, QuickStatsBar and FilterBar
 
 ### Changed
 
+- Refactor InventorySection to accept props and streamline rendering
 - Refactor ItemCard and InventorySection to render forms and additional components.
 - Updated App tests to validate componentized rendering (header, navigation, inventory sections, and footer) instead of only a basic smoke check.